requirements.txt
bs4
pymediainfo
gtts
pydub
cinemagoer
edge-tts
emoji
geopy
git-python
glitch_this
googlesearch-python
googletrans @ git+https://gitlab.com/badmunda98/py-googletrans
heroku3
httpx
lottie
lyricsgenius
motor
pillow
psutil
pyroaddon
pySmartDL
python-dotenv
pytz
requests
selenium
speedtest-cli
telegraph
tgcrypto
Unidecode
wikipedia
py-tgcalls
yt-dlp
kurigram>=2.2.13
